src/main/java/Q00468m.java


/**
 * 编写一个函数来验证输入的字符串是否是有效的 IPv4 或IPv6 地址。
 *
 * 如果是有效的 IPv4 地址，返回 "IPv4" ；
 * 如果是有效的 IPv6 地址，返回 "IPv6" ；
 * 如果不是上述类型的 IP 地址，返回 "Neither" 。
 * IPv4地址由十进制数和点来表示，每个地址包含 4 个十进制数，其范围为0 -255，用(".")分割。比如，172.16.254.1；
 *
 * 同时，IPv4 地址内的数不会以 0 开头。比如，地址172.16.254.01 是不合法的。
 *
 * IPv6地址由 8 组 16 进制的数字来表示，每组表示16 比特。这些组数字通过 (":")分割。比如,2001:0db8:85a3:0000:0000:8a2e:0370:7334 是一个有效的地址。而且，我们可以加入一些以 0 开头的数字，字母可以使用大写，也可以是小写。所以，2001:db8:85a3:0:0:8A2E:0370:7334 也是一个有效的 IPv6 address地址 (即，忽略 0 开头，忽略大小写)。
 *
 * 然而，我们不能因为某个组的值为 0，而使用一个空的组，以至于出现 (::) 的情况。比如，2001:0db8:85a3::8A2E:0370:7334 是无效的 IPv6 地址。
 *
 * 同时，在 IPv6 地址中，多余的 0 也是不被允许的。比如，02001:0db8:85a3:0000:0000:8a2e:0370:7334 是无效的。
 *
 * 示例 1：
 * 输入：IP = "172.16.254.1"
 * 输出："IPv4"
 * 解释：有效的 IPv4 地址，返回 "IPv4"
 *
 * 示例 2：
 * 输入：IP = "2001:0db8:85a3:0:0:8A2E:0370:7334"
 * 输出："IPv6"
 * 解释：有效的 IPv6 地址，返回 "IPv6"
 *
 * 示例 3：
 * 输入：IP = "256.256.256.256"
 * 输出："Neither"
 * 解释：既不是 IPv4 地址，又不是 IPv6 地址
 *
 * 示例 4：
 * 输入：IP = "2001:0db8:85a3:0:0:8A2E:0370:7334:"
 * 输出："Neither"
 *
 * 示例 5：
 * 输入：IP = "1e1.4.5.6"
 * 输出："Neither"
 *
 * 提示：
 * IP 仅由英文字母，数字，字符 '.' 和 ':' 组成。
 *
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/validate-ip-address
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 */
public class Q00468m {

    private static final String Neither = "Neither";
    private static final String IPv4 = "IPv4";
    private static final String IPv6 = "IPv6";

    public String validIPAddress(String IP) {
        if (IP == null || IP.length() < 7) return Neither;
        String[] arr = IP.split("\\.", -1); // limit=-1不会去掉结尾的N个.
        if (arr.length != 4) { // 不是IPv4
            arr = IP.split(":", -1); // limit=-1不会去掉结尾的N个:
            if (arr.length != 8) { // 也不是IPv6
                return Neither;
            }
            // 可能是IPv6
            for (String s : arr) {
                if (s.length() == 0 || s.length() > 4) return Neither;
                for (int i = 0; i < s.length(); i++) {
                    if (!Character.isDigit(s.charAt(i)) && !isLetter(s.charAt(i))) return Neither;
                }
            }
            return IPv6;
        } else { // 可能是IPv4
            // 校验.号个数
            for (String s : arr) {
                if (s.length() == 0 || s.length() > 3 || (s.charAt(0) == '0' && s.length() != 1)) return Neither;
                for (int i = 0; i < s.length(); i++) {
                    if (!Character.isDigit(s.charAt(i))) return Neither;
                }
                int num = Integer.parseInt(s);
                if (num > 255) return Neither;
            }
            return IPv4;
        }
    }

    private boolean isLetter(char c) {
        return c == 'A' || c == 'a' ||
                c == 'B' || c == 'b' ||
                c == 'C' || c == 'c' ||
                c == 'D' || c == 'd' ||
                c == 'E' || c == 'e' ||
                c == 'F' || c == 'f';
    }
}
